requirements.txt
Django~=2.2.4
django-debug-toolbar~=3.2